% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.app.end}
\alias{visr.app.end}
\title{End app definition}
\usage{
visr.app.end(printjson = FALSE, writefile = FALSE, filename = NULL)
}
\arguments{
\item{printjson}{whether to print the generated json file to console}

\item{writefile}{whether to write the generated json to a file}

\item{filename}{path to the filename to write the json to. If not specified and
writefile is TRUE, a json file is generated from the caller
source file path by replacing .R with .json}
}
\description{
Finishes the current apps parameter definition.
}

