% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.app.start}
\alias{visr.app.start}
\title{Start app definition}
\usage{
visr.app.start(name, info = "", debugdata = NULL)
}
\arguments{
\item{name}{app name}

\item{info}{app info shown as tooltip}

\item{debugdata}{debug dataframe to be used when debuggin the app in R / RStudio}
}
\description{
Starts definition of parameters for an R app.
}
\examples{
visr.app.start("kmeans", info="kmeans clustering")
visr.app.start("kmeans", debugdata = iris) # will assign visr.input <- iris in debug mode
}

