% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.applyParameters}
\alias{visr.applyParameters}
\title{Apply GUI parameters}
\usage{
visr.applyParameters()
}
\description{
Applies (imports) parameters spcified in the visrseq GUI to the R environment.
}

