% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.biocLite}
\alias{visr.biocLite}
\title{Load/install app dependent package}
\usage{
visr.biocLite(pkg)
}
\arguments{
\item{pkg}{package name}
}
\description{
Loads a bioconductor package. If not already installed, tries to install the package from bioconductor.
}

