% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.category}
\alias{visr.category}
\title{Start category}
\usage{
visr.category(label, info = "")
}
\arguments{
\item{label}{category label to be shown in VisRseq}

\item{info}{additional information about category shown as tooltip}
}
\description{
Starts a new category of parameters for the app
}
\examples{
visr.category("clustering parameters")
}

