% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.isGUI}
\alias{visr.isGUI}
\title{Check if running in GUI (VisRseq)}
\usage{
visr.isGUI()
}
\value{
Will return \code{TRUE} if the code is running from within VisRseq.
}
\description{
Checks if running from within the VisRseq.
}

