% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.message}
\alias{visr.message}
\title{Show message dialog}
\usage{
visr.message(text, type = c("error", "warning"))
}
\arguments{
\item{text}{Message text to be shown.}

\item{type}{Message type. (\code{"error"} or \code{"warning"})}
}
\description{
Shows a message dialog to the user in VisRseq.
}
\examples{
if (any(is.na(visr.input)))
    visr.message("There are NA values in the input", type="error")
}

