% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visrutils.R
\name{visr.param}
\alias{visr.param}
\title{Add app parameter}
\usage{
visr.param(name, label = NULL, info = NULL, type = c("string",
  "character", "int", "integer", "double", "boolean", "logical", "multi-string",
  "column", "multi-column", "column-numerical", "multi-column-numerical",
  "color", "multi-color", "filename", "output-column", "output-multi-column",
  "output-table"), default = NULL, min = NULL, max = NULL, items = NULL,
  item.labels = NULL, filename.mode = c("load", "save", "dir"),
  debugvalue = NULL)
}
\arguments{
\item{name}{Parameter name. will be appended to "visr.param." to create
the full variable name in R.}

\item{label}{The label for the parameter's GUI control.
Will use variable name if not specified (NULL).}

\item{info}{Additional information about parameter.
Will be shown as tooltip in VisRseq.}

\item{type}{Parameter type}

\item{default}{Initial default value for the parameter.}

\item{min}{Minimum value for numerical type parameters
(\code{"int"}, \code{"double"})}

\item{max}{Maximum value for numerical type parameters
(\code{"int"}, \code{"double"})}

\item{items}{Specify a vector of items for a \code{"string"} variable to select from.}

\item{item.labels}{Specify a vector of items to be used as the labels for \code{items} argument.}

\item{filename.mode}{Specify the file dialog mode
(file load, file save or directory) for a
\code{"filename"} type parameter}

\item{debugvalue}{The value to be assigned to the R variable in debug mode.
Useful for unit testing.}
}
\description{
Adds a new parameter to the curent app.
}
\examples{
visr.param("k", type = "integer") # specify type
visr.param("k", default = 3L) # will infer type from default value
visr.param("k", label = "number of clusters")  # explicitly specify label
visr.param("title") # no type or default value: treated as a "string" type
visr.param("algorithm", items = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"))
visr.param("columns", type = "multi-column-numerical")
visr.param("output.clusterid", type = "output-column") # column appended to input table
}

